+++ src/features/welcome/components/Counter/index.tsx
@@ -1,0 +1,1 @@
+export { Counter as default } from './Counter'
