+++ src/features/welcome/components/Counter/style.module.scss
@@ -1,0 +1,3 @@
+.buttonRow {
+  margin: 0;
+}
